src/requirements.txt
sendgrid
PyGithub
pyyaml
flask
